+++ src/core/create/resolveSelection.ts
@@ -1,5 +1,5 @@
 import * as vs from "vscode"
-import { LogInfo } from "."
+import { LogInfo, Options } from "."
 import { createLogInfo } from "../helper"
 import {
   generateLog,
@@ -9,12 +9,9 @@ import {
   isFunction,
 } from "../../utils"
 
-export function resolveSelection(
-  editor: vs.TextEditor,
-  consoleInObject: boolean | undefined
-): LogInfo {
+export function resolveSelection(editor: vs.TextEditor, options: Options): LogInfo {
   const logInfo = createLogInfo()
-  getLogsAndCursor(editor, logInfo, consoleInObject)
+  getLogsAndCursor(editor, logInfo, options)
 
   return logInfo
 }
@@ -22,7 +19,7 @@ export function resolveSelection(
 function getLogsAndCursor(
   editor: vs.TextEditor,
   logInfo: LogInfo & { push: (log: string) => void },
-  consoleInObject: boolean | undefined
+  options: Options
 ): void {
   const document = editor.document
   const selectedText = document.getText(editor.selection)
@@ -30,17 +27,18 @@ function getLogsAndCursor(
   const lineText = getLineText(document, line)
   const words = getWordsFromSelected(selectedText, lineText)
   const { cursorPosition, push } = logInfo
+  const { consoleInObject, consoleVariablesName } = options
 
   const space = getStartSpace(lineText)
   if (consoleInObject) {
-    push(generateLogInObject(words, space))
+    push(generateLogInObject(words, space, consoleVariablesName))
     cursorPosition.line = line + 1
     cursorPosition.character = logInfo.logs.length - 1
   } else {
     for (let i = 0; i < words.length - 1; i++) {
-      push(generateLog(words[i], space))
+      push(generateLog(words[i], space, consoleVariablesName))
     }
-    const lastLog = generateLog(words[words.length - 1], space)
+    const lastLog = generateLog(words[words.length - 1], space, consoleVariablesName)
     push(lastLog)
 
     cursorPosition.line = line + words.length