+++ src/extension.ts
@@ -5,8 +5,6 @@ export function activate(context: vs.ExtensionContext) {
   const createCommandId = "quickConsole.createConsoleLog"
   const clearCommandId = "quickConsole.clearConsoleLog"
 
-  console.log(vs.workspace.getConfiguration("Quick Console"))
-
   const create = vs.commands.registerCommand(createCommandId, createConsoleLog)
   const clear = vs.commands.registerCommand(clearCommandId, clearConsoleLog)